reiptables: Add UDP protocol type bit to divert packet mark
It might happen that UDP protocol could be diverted to local application
working as transparent proxy (e.g. DNS) and there is no way to determine
from packet mark if port stored in divert id field of packet mark is TCP
or UDP.

Use bit 17 from reserved bits range for this purpose, keep it as zero if
protocol is TCP and set to 1 if it is UDP.

There is no functional changes here as we neither have UDP rules nor
inspect reserved bits anywhere in rules.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>